v1.8.3
added a method for scaling images to the stepwindow ui

reworked nearly all files to make include of qtdatavisualization possible. somthing in there was abitous with something in namespace cv. So all instances of "using namespace cv" were removed. Now there ist a header file d_opencv_typedefs.h" with a lot of typedefs and enum redefinition from opencv. that is included whenever opencv stuff is needed. That way the code does not need to be changed with a ton of "cv::".